clustermesh: move service export sync to cm-apiserver pkg
Similarly to the previous commit also move service export sync from the
mcs-api package to clustermesh-apiserver, this was the last resource not
already in that package!

Also transition from the existing unit test to hive scripts to match the
other resource and modernize the testing of this resource syncing.
