+++ clustermesh-apiserver/clustermesh/testdata/globalnamespace-serviceexports.txtar
@@ -1,0 +1,181 @@
+#! --cluster-id=3 --cluster-name=cluster3 --clustermesh-enable-mcs-api --clustermesh-default-global-namespace=false
+
+# Create ServiceExport CRD to pass CRD existence validation
+k8s/add serviceexport-crd.yaml
+
+hive/start
+
+# Add two namespaces - ns-alpha (annotated as global), ns-beta (not global)
+k8s/add namespace-alpha.yaml namespace-beta.yaml
+
+# Add two service exports - one in each namespace
+k8s/add service-alpha.yaml service-beta.yaml
+k8s/add serviceexport-alpha.yaml serviceexport-beta.yaml
+
+# Assert that the synced key gets created
+kvstore/list -o plain cilium/synced synced.actual
+* grep -q '^# cilium/synced/cluster3/cilium/state/serviceexports/v1$' synced.actual
+
+# Wait for synchronization. Only service export in ns-alpha should be synced (global namespace)
+# The service export in ns-beta should NOT be synced because its namespace is not global
+kvstore/list -o json cilium/state/serviceexports serviceexports-1.actual
+* cmp serviceexports-1.actual serviceexports-1.expected
+
+# Annotate ns-beta as global
+k8s/update namespace-beta-annotated.yaml
+
+# Wait for synchronization. Now both service exports should be synced
+kvstore/list -o json cilium/state/serviceexports serviceexports-1+2.actual
+* cmp serviceexports-1+2.actual serviceexports-1+2.expected
+
+# Remove global annotation from ns-beta
+k8s/update namespace-beta.yaml
+
+# Wait for synchronization. Only service export in ns-alpha should remain synced
+kvstore/list -o json cilium/state/serviceexports serviceexports-1.actual
+* cmp serviceexports-1.actual serviceexports-1.expected
+
+# Delete the service export in ns-alpha
+k8s/delete serviceexport-alpha.yaml
+
+# Wait for synchronization. No service exports should remain synced
+kvstore/list -o json cilium/state/serviceexports serviceexports-empty.actual
+* empty serviceexports-empty.actual
+
+# ---
+
+-- serviceexport-crd.yaml --
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: serviceexports.multicluster.x-k8s.io
+spec:
+  group: multicluster.x-k8s.io
+  names:
+    kind: ServiceExport
+    plural: serviceexports
+    singular: serviceexport
+  scope: Namespaced
+  versions:
+  - name: v1beta1
+    served: true
+    storage: true
+
+-- namespace-alpha.yaml --
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: ns-alpha
+  annotations:
+    clustermesh.cilium.io/global: "true"
+
+-- namespace-beta.yaml --
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: ns-beta
+
+-- namespace-beta-annotated.yaml --
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: ns-beta
+  annotations:
+    clustermesh.cilium.io/global: "true"
+
+-- service-alpha.yaml --
+apiVersion: v1
+kind: Service
+metadata:
+  name: echo-service
+  namespace: ns-alpha
+spec:
+  ports:
+  - name: http
+    port: 80
+    protocol: TCP
+  selector:
+    app: echo
+  type: ClusterIP
+
+-- service-beta.yaml --
+apiVersion: v1
+kind: Service
+metadata:
+  name: echo-service
+  namespace: ns-beta
+spec:
+  ports:
+  - name: http
+    port: 80
+    protocol: TCP
+  selector:
+    app: echo
+  type: ClusterIP
+
+-- serviceexport-alpha.yaml --
+apiVersion: multicluster.x-k8s.io/v1beta1
+kind: ServiceExport
+metadata:
+  name: echo-service
+  namespace: ns-alpha
+  creationTimestamp: "2026-01-01T00:00:00Z"
+
+-- serviceexport-beta.yaml --
+apiVersion: multicluster.x-k8s.io/v1beta1
+kind: ServiceExport
+metadata:
+  name: echo-service
+  namespace: ns-beta
+  creationTimestamp: "2026-01-01T00:00:00Z"
+
+-- serviceexports-1.expected --
+# cilium/state/serviceexports/v1/cluster3/ns-alpha/echo-service
+{
+  "cluster": "cluster3",
+  "name": "echo-service",
+  "namespace": "ns-alpha",
+  "exportCreationTimestamp": "2026-01-01T00:00:00Z",
+  "ports": [
+    {
+      "name": "http",
+      "protocol": "TCP",
+      "port": 80
+    }
+  ],
+  "type": "ClusterSetIP",
+  "sessionAffinity": ""
+}
+-- serviceexports-1+2.expected --
+# cilium/state/serviceexports/v1/cluster3/ns-alpha/echo-service
+{
+  "cluster": "cluster3",
+  "name": "echo-service",
+  "namespace": "ns-alpha",
+  "exportCreationTimestamp": "2026-01-01T00:00:00Z",
+  "ports": [
+    {
+      "name": "http",
+      "protocol": "TCP",
+      "port": 80
+    }
+  ],
+  "type": "ClusterSetIP",
+  "sessionAffinity": ""
+}
+# cilium/state/serviceexports/v1/cluster3/ns-beta/echo-service
+{
+  "cluster": "cluster3",
+  "name": "echo-service",
+  "namespace": "ns-beta",
+  "exportCreationTimestamp": "2026-01-01T00:00:00Z",
+  "ports": [
+    {
+      "name": "http",
+      "protocol": "TCP",
+      "port": 80
+    }
+  ],
+  "type": "ClusterSetIP",
+  "sessionAffinity": ""
+}
