+++ clustermesh-apiserver/clustermesh/testdata/services.txtar
